Reduce reliance on Alloy's U256 type

+++ substrate/frame/revive/fixtures/contracts/TwoSlots.sol
@@ -2,12 +2,14 @@
 pragma solidity ^0.8.4;
 
 /// Minimal two-slot storage contract for testing storage overrides.
-/// `first` occupies slot 0, `second` occupies slot 1.
+/// Each variable is declared as `uint256` to guarantee separate storage slots
+/// (slot 0 for `first`, slot 1 for `second`), while the constructor accepts
+/// `uint64` to keep the ABI lightweight.
 contract TwoSlots {
     uint256 public first;
     uint256 public second;
 
-    constructor(uint256 _first, uint256 _second) {
+    constructor(uint64 _first, uint64 _second) {
         first = _first;
         second = _second;
     }
